remove -Wl,--unresolved-symbols=ignore-in-object-files from CFLAGS
There's actually quite some errors when building for freestanding.
To fix them, the configure script has to be patched to include LIBS in
the conftest. Also, we can only build the `mpn`, `mpz`, `mpq` and `mpf`
folders. For example printing functions require `islower` and `isupper`
but they are not available in ocaml-freestanding.